fix(symthaea): ort 2.0 API, clippy, cargo-deny v2, CI-safe integration tests
- kokoro_engine: commit_from_file + Tensor::from_array (ort 2.0 compat)
- Cargo.toml: add `parallel` feature for WASM-compat rayon gating
- cycle_phase_dynamics: remove unnecessary f32 cast (clippy)
- deny.toml: remove deprecated advisories fields (cargo-deny v2)
- e2e_pipeline + pathology: graceful skip on CI permissions/LLM errors

+++ src/cognitive_loop/cycle_phase_dynamics.rs
@@ -546,7 +546,7 @@ impl CognitiveLoopService {
         // Science: McClelland et al. (1995) — complementary learning systems.
         if resonator_best_sim > RESONATOR_CONSOLIDATION_THRESHOLD {
             self.fep.agent.precision.prior_precision = (self.fep.agent.precision.prior_precision
-                + (resonator_best_sim - RESONATOR_CONSOLIDATION_THRESHOLD as f32) as f64 * 0.1)
+                + (resonator_best_sim - RESONATOR_CONSOLIDATION_THRESHOLD) as f64 * 0.1)
                 .min(2.0);
             if self.stats.total_cycles > 10 {
                 self.scale_lr("resonator_familiar", RESONATOR_FAMILIAR_LR_SCALE);
